Improve fern behavior
It just feels more vim-like, and works a bit better imo to do this in
the same window that you call :Fern from. Will be more jarring for
anyone coming from a VS Code / Atom / Sublime background etc but I think
it's worth it.

The gist of it is: you call :Fern from the pane that you want to open
the new file in.